+++ src/gmi/mostusedsites/backend/tests/test_models.py
@@ -20,13 +20,13 @@ class TestModels:
             commit()
 
     def test_visit(self, session):
-        visit = Visit(url='http://foo', visited_at=1, duration=1, active=True)
+        visit = Visit('http://foo', visited_at=1, duration=1, active=True)
         session.add(visit)
         assert(commit() == None)
 
     def test_visit_belongs_to_user(self, session):
         user = User(unique_id='foo')
-        visit = Visit(url='http://foo', visited_at=1, duration=1, active=True)
+        visit = Visit('http://foo', visited_at=1, duration=1, active=True)
         user.visits.append(visit)
 
         assert(visit.user == user)
@@ -36,32 +36,32 @@ class TestModels:
             visit = Visit(visited_at=1, duration=1, active=True)
 
     def test_visit_visited_at_is_present(self, session):
-        visit = Visit(url='http://foo_visited', duration=1, active=True)
+        visit = Visit('http://foo_visited', duration=1, active=True)
         session.add(visit)
 
         with pytest.raises(IntegrityError):
             commit()
 
     def test_visit_duration_is_present(self, session):
-        visit = Visit(url='http://foo_duration', visited_at=1, active=True)
+        visit = Visit('http://foo_duration', visited_at=1, active=True)
         session.add(visit)
 
         with pytest.raises(IntegrityError):
             commit()
 
     def test_visit_active_is_present(self, session):
-        visit = Visit(url='http://foo_active', duration=1, visited_at=1)
+        visit = Visit('http://foo_active', duration=1, visited_at=1)
         session.add(visit)
 
         with pytest.raises(IntegrityError):
             commit()
 
     def test_visit_id(self, session):
-        visit = Visit(url='http://foo', visited_at=1, duration=1)
+        visit = Visit('http://foo', visited_at=1, duration=1)
         assert(visit.id == '1aa0598ba12ba82e6b6a88f97f010948f33a01d5')
 
     def test_visit_id_unique(self, session):
-        visit = Visit(url='http://bar', visited_at=1, duration=1)
+        visit = Visit('http://bar', visited_at=1, duration=1)
         visit1 = Visit(url='http://bar', visited_at=1, duration=1)
         session.add(visit)
         session.add(visit1)
@@ -70,21 +70,21 @@ class TestModels:
             commit()
 
     def test_visit_id_long_timestamp(self, session):
-        visit = Visit(url="http://bar", visited_at=1449073100894, duration=1)
+        visit = Visit("http://bar", visited_at=1449073100894, duration=1)
         assert(visit.id == 'cdd896d38b96f973e2d0f446daf5da090878f585')
 
     def test_visit_schema(self, session):
-        visit = Visit(url='https://bar', visited_at=1, duration=1)
+        visit = Visit('https://bar', visited_at=1, duration=1)
         assert(visit.scheme == 'https')
 
     def test_visit_host(self, session):
-        visit = Visit(url='https://bar', visited_at=1, duration=1)
+        visit = Visit('https://bar', visited_at=1, duration=1)
         assert(visit.host == 'bar')
 
     def test_visit_path(self, session):
-        visit = Visit(url='https://bar/foo/baz', visited_at=1, duration=1)
+        visit = Visit('https://bar/foo/baz', visited_at=1, duration=1)
         assert(visit.path == '/foo/baz')
 
     def test_visit_no_params(self, session):
-        visit = Visit(url='https://bar/foo?a=1&b=2', visited_at=1, duration=1)
+        visit = Visit('https://bar/foo?a=1&b=2', visited_at=1, duration=1)
         assert(visit.path == '/foo')