+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog for "Most used sites" backend
 - JSON responses, split url in host, scheme and path.
 - JSON responses, only expose host when requesting third party visits for
   privacy reasons.
+- Remove url from visit model.
 
 
 0.1 (2015-12-08)